тут куча всего. Но главное - это гардероб фулл рабочий.

+++ app/src/main/java/com/example/mobileapp/models/auth/RegisterResponse.kt
@@ -1,6 +1,6 @@
 package com.example.mobileapp.models.auth
 
 data class RegisterResponse(
-    val userId: Int? = null,
+    val userId: Long? = null,
     val error: String? = null
 )
